REF: test_rename_axis (#33152)
