+++ pandas/tests/series/test_alter_axes.py
@@ -3,7 +3,7 @@ from datetime import datetime
 import numpy as np
 import pytest
 
-from pandas import Index, MultiIndex, Series
+from pandas import Index, Series
 import pandas._testing as tm
 
 
@@ -53,43 +53,6 @@ class TestSeriesAlterAxes:
         s = Series(range(10))
         s.index = idx
         assert s.index.is_all_dates
-
-    def test_rename_axis_mapper(self):
-        # GH 19978
-        mi = MultiIndex.from_product([["a", "b", "c"], [1, 2]], names=["ll", "nn"])
-        s = Series(list(range(len(mi))), index=mi)
-
-        result = s.rename_axis(index={"ll": "foo"})
-        assert result.index.names == ["foo", "nn"]
-
-        result = s.rename_axis(index=str.upper, axis=0)
-        assert result.index.names == ["LL", "NN"]
-
-        result = s.rename_axis(index=["foo", "goo"])
-        assert result.index.names == ["foo", "goo"]
-
-        with pytest.raises(TypeError, match="unexpected"):
-            s.rename_axis(columns="wrong")
-
-    def test_rename_axis_inplace(self, datetime_series):
-        # GH 15704
-        expected = datetime_series.rename_axis("foo")
-        result = datetime_series
-        no_return = result.rename_axis("foo", inplace=True)
-
-        assert no_return is None
-        tm.assert_series_equal(result, expected)
-
-    @pytest.mark.parametrize("kwargs", [{"mapper": None}, {"index": None}, {}])
-    def test_rename_axis_none(self, kwargs):
-        # GH 25034
-        index = Index(list("abc"), name="foo")
-        df = Series([1, 2, 3], index=index)
-
-        result = df.rename_axis(**kwargs)
-        expected_index = index.rename(None) if kwargs else index
-        expected = Series([1, 2, 3], index=expected_index)
-        tm.assert_series_equal(result, expected)
 
     def test_set_axis_inplace_axes(self, axis_series):
         # GH14636
